types: Move Array to array.py
